rename: MenuId.ts → PageId.ts
