+++ types/index.ts
@@ -1,3 +1,3 @@
 export type { LayoutProps } from "./LayoutProps";
 export type { IconProps } from "./IconProps";
-export type { MenuId } from "./MenuId";
+export type { MenuId } from "./PageId";
